wip
- Convert to Kotlin style Gradle
- Change namespace to no.novari.flyt.kafka
- Bump to latest Kafka RC
- Published as `no.novari:flyt-kafka-{version}`
