--- build.gradle
@@ -1,62 +1,0 @@
-import org.springframework.boot.gradle.plugin.SpringBootPlugin
-
-plugins {
-    id 'org.springframework.boot' version '3.5.6' apply false
-    id 'io.spring.dependency-management' version '1.1.7'
-    id 'java-library'
-    id 'maven-publish'
-}
-
-group = 'no.fintlabs'
-version = findProperty("version") ?: "1.0-SNAPSHOT"
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(21)
-    }
-    withSourcesJar()
-}
-
-publishing {
-    repositories {
-        mavenLocal()
-    }
-}
-
-repositories {
-    mavenLocal()
-    maven {
-        url = "https://repo.fintlabs.no/releases"
-    }
-    mavenCentral()
-}
-
-dependencyManagement {
-    imports {
-        mavenBom SpringBootPlugin.BOM_COORDINATES
-    }
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-
-    implementation 'org.springframework.kafka:spring-kafka'
-    implementation 'no.fintlabs:fint-kafka:5.0.0-rc-12'
-    testImplementation 'org.springframework.kafka:spring-kafka-test'
-
-    compileOnly 'org.projectlombok:lombok'
-    runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
-    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
-    annotationProcessor 'org.projectlombok:lombok'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'io.projectreactor:reactor-test'
-    testImplementation 'cglib:cglib-nodep:3.3.0'
-}
-
-test {
-    useJUnitPlatform()
-}
-
-apply from: 'https://raw.githubusercontent.com/FINTLabs/fint-buildscripts/master/reposilite.ga.gradle'
