+++ build.gradle.kts
@@ -1,0 +1,80 @@
+import org.springframework.boot.gradle.plugin.SpringBootPlugin
+
+plugins {
+    id("org.springframework.boot") version "3.5.7" apply false
+    id("io.spring.dependency-management") version "1.1.7"
+    id("java-library")
+    id("maven-publish")
+    id("com.github.ben-manes.versions") version "0.53.0"
+}
+
+group = "no.novari"
+version = findProperty("version") as String? ?: "1.0-SNAPSHOT"
+
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(21))
+    }
+    withSourcesJar()
+}
+
+publishing {
+    repositories {
+        mavenLocal()
+    }
+}
+
+repositories {
+    mavenLocal()
+    maven {
+        url = uri("https://repo.fintlabs.no/releases")
+    }
+    mavenCentral()
+}
+
+dependencyManagement {
+    imports {
+        mavenBom(SpringBootPlugin.BOM_COORDINATES)
+    }
+}
+
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-actuator")
+    implementation("org.springframework.boot:spring-boot-starter-webflux")
+
+    implementation("org.springframework.kafka:spring-kafka")
+    implementation("no.novari:kafka:5.0.0-rc-16")
+    testImplementation("org.springframework.kafka:spring-kafka-test")
+
+    compileOnly("org.projectlombok:lombok")
+    runtimeOnly("io.micrometer:micrometer-registry-prometheus")
+    annotationProcessor("org.springframework.boot:spring-boot-configuration-processor")
+    annotationProcessor("org.projectlombok:lombok")
+
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("io.projectreactor:reactor-test")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+publishing {
+    repositories {
+        maven {
+            url = uri("https://repo.fintlabs.no/releases")
+            credentials {
+                username = System.getenv("REPOSILITE_USERNAME")
+                password = System.getenv("REPOSILITE_PASSWORD")
+            }
+            authentication {
+                create<BasicAuthentication>("basic")
+            }
+        }
+    }
+    publications {
+        create<MavenPublication>("maven") {
+            from(components["java"])
+        }
+    }
+}
