+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'fint-flyt-kafka'
+rootProject.name = 'flyt-kafka'
