--- src/main/java/no/fintlabs/flyt/kafka/instanceflow/headers/CouldNotReadInstanceFlowHeadersException.java
@@ -1,11 +1,0 @@
-package no.fintlabs.flyt.kafka.instanceflow.headers;
-
-import org.apache.kafka.common.header.Header;
-
-public class CouldNotReadInstanceFlowHeadersException extends RuntimeException {
-
-    public CouldNotReadInstanceFlowHeadersException(Header header) {
-        super("Could not read instance flow headers from " + header);
-    }
-
-}
