--- src/main/java/no/fintlabs/flyt/kafka/instanceflow/headers/CouldNotWriteInstanceFlowHeadersException.java
@@ -1,9 +1,0 @@
-package no.fintlabs.flyt.kafka.instanceflow.headers;
-
-public class CouldNotWriteInstanceFlowHeadersException extends RuntimeException {
-
-    public CouldNotWriteInstanceFlowHeadersException(InstanceFlowHeaders instanceFlowHeaders) {
-        super("Could not write " + instanceFlowHeaders + "to bytes");
-    }
-
-}
