--- src/main/java/no/fintlabs/flyt/kafka/instanceflow/headers/NoInstanceFlowHeadersException.java
@@ -1,8 +1,0 @@
-package no.fintlabs.flyt.kafka.instanceflow.headers;
-
-public class NoInstanceFlowHeadersException extends RuntimeException {
-
-    public NoInstanceFlowHeadersException() {
-        super("No instance flow headers");
-    }
-}
