--- src/main/java/no/fintlabs/flyt/kafka/instanceflow/producing/InstanceFlowTemplateFactory.java
@@ -1,27 +1,0 @@
-package no.fintlabs.flyt.kafka.instanceflow.producing;
-
-import no.fintlabs.flyt.kafka.instanceflow.headers.InstanceFlowHeadersMapper;
-import no.fintlabs.kafka.producing.ParameterizedTemplateFactory;
-import org.springframework.stereotype.Service;
-
-@Service
-public class InstanceFlowTemplateFactory {
-
-    private final ParameterizedTemplateFactory parameterizedTemplateFactory;
-    private final InstanceFlowHeadersMapper instanceFlowHeadersMapper;
-
-    public InstanceFlowTemplateFactory(
-            ParameterizedTemplateFactory parameterizedTemplateFactory,
-            InstanceFlowHeadersMapper instanceFlowHeadersMapper
-    ) {
-        this.parameterizedTemplateFactory = parameterizedTemplateFactory;
-        this.instanceFlowHeadersMapper = instanceFlowHeadersMapper;
-    }
-
-    public <VALUE> InstanceFlowTemplate<VALUE> createTemplate(Class<VALUE> valueClass) {
-        return new InstanceFlowTemplate<>(
-                parameterizedTemplateFactory.createTemplate(valueClass),
-                instanceFlowHeadersMapper
-        );
-    }
-}
