+++ src/main/java/no/novari/flyt/kafka/instanceflow/consuming/InstanceFlowConsumerRecord.java
@@ -1,0 +1,13 @@
+package no.novari.flyt.kafka.instanceflow.consuming;
+
+import lombok.Builder;
+import lombok.Getter;
+import no.novari.flyt.kafka.instanceflow.headers.InstanceFlowHeaders;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+
+@Builder
+@Getter
+public final class InstanceFlowConsumerRecord<T> {
+    private final InstanceFlowHeaders instanceFlowHeaders;
+    private final ConsumerRecord<String, T> consumerRecord;
+}
