+++ src/main/java/no/novari/flyt/kafka/instanceflow/consuming/InstanceFlowErrorHandlerConfigurationStepBuilder.java
@@ -1,0 +1,186 @@
+package no.novari.flyt.kafka.instanceflow.consuming;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import org.apache.kafka.clients.consumer.Consumer;
+import org.apache.logging.log4j.util.TriConsumer;
+import org.springframework.kafka.support.ExponentialBackOffWithMaxRetries;
+import org.springframework.util.backoff.BackOff;
+import org.springframework.util.backoff.ExponentialBackOff;
+import org.springframework.util.backoff.FixedBackOff;
+
+import java.time.Duration;
+import java.util.Optional;
+import java.util.function.BiConsumer;
+import java.util.function.BiFunction;
+
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class InstanceFlowErrorHandlerConfigurationStepBuilder {
+
+    static <VALUE> InstanceFlowErrorHandlerConfigurationStepBuilder.RetryStep<VALUE> firstStep(Class<VALUE> consumerRecordValueClass) {
+        return new InstanceFlowErrorHandlerConfigurationStepBuilder.Steps<>(consumerRecordValueClass);
+    }
+
+    public interface RetryStep<VALUE> extends InstanceFlowErrorHandlerConfigurationStepBuilder.DefaultRetryStep<VALUE>, InstanceFlowErrorHandlerConfigurationStepBuilder.RetryFunctionStep<VALUE> {
+    }
+
+    public interface DefaultRetryStep<VALUE> {
+        RecoveryStep<VALUE> noRetries();
+
+        RecoveryStep<VALUE> retryWithFixedInterval(
+                Duration interval,
+                int maxRetries
+        );
+
+        RecoveryStep<VALUE> retryWithExponentialInterval(
+                Duration initialInterval,
+                double intervalMultiplier,
+                Duration maxInterval,
+                Duration maxElapsedTime
+        );
+
+        RecoveryStep<VALUE> retryWithExponentialInterval(
+                Duration initialInterval,
+                double intervalMultiplier,
+                Duration maxInterval,
+                int maxRetries
+        );
+
+    }
+
+    public interface RetryFunctionStep<VALUE> {
+        RetryFunctionDefaultStep<VALUE> retryWithBackoffFunction(
+                BiFunction<InstanceFlowConsumerRecord<VALUE>, Exception, Optional<BackOff>> backoffFunction
+        );
+    }
+
+    public interface RetryFunctionDefaultStep<VALUE> {
+        DefaultRetryStep<VALUE> orElse();
+    }
+
+    public interface RecoveryStep<VALUE> {
+
+        BuilderStep<VALUE> skipFailedRecords();
+
+        BuilderStep<VALUE> handleFailedRecords(
+                TriConsumer<InstanceFlowConsumerRecord<VALUE>, Consumer<String, VALUE>, Exception> customRecoverer
+        );
+
+        BuilderStep<VALUE> handleFailedRecords(
+                BiConsumer<InstanceFlowConsumerRecord<VALUE>, Exception> customRecoverer
+        );
+    }
+
+    public interface BuilderStep<VALUE> {
+        InstanceFlowErrorHandlerConfiguration<VALUE> build();
+    }
+
+    private static class Steps<VALUE> implements
+            RetryStep<VALUE>,
+            DefaultRetryStep<VALUE>,
+            RetryFunctionDefaultStep<VALUE>,
+            RecoveryStep<VALUE>,
+            BuilderStep<VALUE> {
+
+        private final Class<VALUE> consumerRecordValueClass;
+        private BackOff defaultBackOff;
+        private BiFunction<InstanceFlowConsumerRecord<VALUE>, Exception, Optional<BackOff>> backOffFunction;
+        private TriConsumer<InstanceFlowConsumerRecord<VALUE>, Consumer<String, VALUE>, Exception> customRecoverer;
+
+        public Steps(Class<VALUE> consumerRecordValueClass) {
+            this.consumerRecordValueClass = consumerRecordValueClass;
+        }
+
+        @Override
+        public RecoveryStep<VALUE> noRetries() {
+            return this;
+        }
+
+        @Override
+        public RecoveryStep<VALUE> retryWithFixedInterval(Duration interval, int maxRetries) {
+            defaultBackOff = new FixedBackOff(interval.toMillis(), maxRetries);
+            return this;
+        }
+
+        @Override
+        public RecoveryStep<VALUE> retryWithExponentialInterval(
+                Duration initialInterval,
+                double intervalMultiplier,
+                Duration maxInterval,
+                Duration maxElapsedTime
+        ) {
+            ExponentialBackOff exponentialBackOff = new ExponentialBackOff(
+                    initialInterval.toMillis(),
+                    intervalMultiplier
+            );
+            exponentialBackOff.setMaxInterval(maxInterval.toMillis());
+            exponentialBackOff.setMaxElapsedTime(maxElapsedTime.toMillis());
+            defaultBackOff = exponentialBackOff;
+            return this;
+        }
+
+        @Override
+        public RecoveryStep<VALUE> retryWithExponentialInterval(
+                Duration initialInterval,
+                double intervalMultiplier,
+                Duration maxInterval,
+                int maxRetries
+        ) {
+            ExponentialBackOffWithMaxRetries exponentialBackOff = new ExponentialBackOffWithMaxRetries(
+                    maxRetries
+            );
+            exponentialBackOff.setInitialInterval(initialInterval.toMillis());
+            exponentialBackOff.setMultiplier(intervalMultiplier);
+            exponentialBackOff.setMaxInterval(maxInterval.toMillis());
+            defaultBackOff = exponentialBackOff;
+            return this;
+        }
+
+        @Override
+        public RetryFunctionDefaultStep<VALUE> retryWithBackoffFunction(
+                BiFunction<InstanceFlowConsumerRecord<VALUE>, Exception, Optional<BackOff>> backoffFunction
+        ) {
+            backOffFunction = backoffFunction;
+            return this;
+        }
+
+        @Override
+        public DefaultRetryStep<VALUE> orElse() {
+            return this;
+        }
+
+        @Override
+        public BuilderStep<VALUE> skipFailedRecords() {
+            return this;
+        }
+
+        @Override
+        public BuilderStep<VALUE> handleFailedRecords(
+                TriConsumer<InstanceFlowConsumerRecord<VALUE>, Consumer<String, VALUE>, Exception> customRecoverer
+        ) {
+            this.customRecoverer = customRecoverer;
+            return this;
+        }
+
+        @Override
+        public BuilderStep<VALUE> handleFailedRecords(
+                BiConsumer<InstanceFlowConsumerRecord<VALUE>, Exception> customRecoverer
+        ) {
+            this.customRecoverer =
+                    (consumerRecord, consumer, exception)
+                            -> customRecoverer.accept(consumerRecord, exception);
+            return this;
+        }
+
+        @Override
+        public InstanceFlowErrorHandlerConfiguration<VALUE> build() {
+            return new InstanceFlowErrorHandlerConfiguration<>(
+                    consumerRecordValueClass,
+                    backOffFunction,
+                    defaultBackOff,
+                    customRecoverer
+            );
+        }
+    }
+
+}
