+++ src/main/java/no/novari/flyt/kafka/instanceflow/headers/CouldNotWriteInstanceFlowHeadersException.java
@@ -1,0 +1,9 @@
+package no.novari.flyt.kafka.instanceflow.headers;
+
+public class CouldNotWriteInstanceFlowHeadersException extends RuntimeException {
+
+    public CouldNotWriteInstanceFlowHeadersException(InstanceFlowHeaders instanceFlowHeaders) {
+        super("Could not write " + instanceFlowHeaders + "to bytes");
+    }
+
+}
