+++ src/main/java/no/novari/flyt/kafka/instanceflow/headers/NoInstanceFlowHeadersException.java
@@ -1,0 +1,8 @@
+package no.novari.flyt.kafka.instanceflow.headers;
+
+public class NoInstanceFlowHeadersException extends RuntimeException {
+
+    public NoInstanceFlowHeadersException() {
+        super("No instance flow headers");
+    }
+}
