+++ src/main/java/no/novari/flyt/kafka/requestreply/InstanceFlowReplyProducerRecord.java
@@ -1,0 +1,12 @@
+package no.novari.flyt.kafka.requestreply;
+
+import lombok.Builder;
+import lombok.Getter;
+import no.novari.flyt.kafka.instanceflow.headers.InstanceFlowHeaders;
+
+@Builder
+@Getter
+public final class InstanceFlowReplyProducerRecord<T> {
+    private final InstanceFlowHeaders instanceFlowHeaders;
+    private final T value;
+}
