--- src/main/resources/application-flyt-kafka.properties
@@ -1,1 +1,0 @@
-fint.kafka.topic.domain-context=flyt
