+++ src/main/resources/application-flyt-kafka.yml
@@ -1,0 +1,4 @@
+novari:
+  kafka:
+    topic:
+      domain-context: flyt
