--- src/test/java/no/fintlabs/flyt/kafka/TestApplication.java
@@ -1,14 +1,0 @@
-package no.fintlabs.flyt.kafka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan("no.fintlabs")
-public class TestApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TestApplication.class, args);
-    }
-}
