+++ src/test/java/no/novari/flyt/kafka/TestApplication.java
@@ -1,0 +1,12 @@
+package no.novari.flyt.kafka;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "no.novari")
+public class TestApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TestApplication.class, args);
+    }
+}
