+++ src/test/resources/application.yml
@@ -5,7 +5,7 @@ spring:
       group-id: test
     bootstrap-servers: ${spring.embedded.kafka.brokers}
 
-fint:
+novari:
   kafka:
     topic:
       org-id: test-org-id
